Update: Remove unused code, cast (char) to (unsigned int) for array indexes, and fix bitwise problem.
A bit of stale code is exposed by running the compiler with -Wall.
Example:
  fake clean build -d -Wall
  fake clean build -d -Wall -m clang

Using char (generally) is fine because the numbers match.
However, there tends to be specific cases and behaviors that might result in char being not treated as expected.
Explicitly cast to an (unsigned int) to play it safe.

The fwrite_unlocked() response checks were previously mass refactored to use a size check on the response.
Mistakes in this resulted in the the not operation "!" being left there resulting in a bad if condition check.

Clang warnings about not having parenthesis when using "&&" and "||".
I would argue that this is simply an ignorance or incompetence in the programmers.
The programmers should be expected to understand basic parts of a language, such as order of operations.
Rather than fight this battle, I am just adding parenthesis.

+++ level_0/f_print/c/print.c
@@ -55,7 +55,7 @@ extern "C" {
       }
     }
     else {
-      if (fwrite_unlocked(f_print_sequence_set_control_s[character].string, 1, f_print_sequence_set_control_s[character].used, output) == f_print_sequence_set_control_s[character].used) {
+      if (fwrite_unlocked(f_print_sequence_set_control_s[(unsigned int) character].string, 1, f_print_sequence_set_control_s[(unsigned int) character].used, output) == f_print_sequence_set_control_s[(unsigned int) character].used) {
         return F_none;
       }
     }
